refactor: uptime component timeline
